+++ wasmbinding/bindings/pool.go
@@ -1,8 +1,0 @@
-package bindings
-
-import sdk "github.com/cosmos/cosmos-sdk/types"
-
-type PoolAssets struct {
-	Assets []sdk.Coin
-	Shares sdk.Coin
-}